LevelSelect

+++ Assets/Script/CoinCollectable.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class CoinCollectable : MonoBehaviour
 {
     [SerializeField]
-    int rotateSpeed = 1;
+    float rotateSpeed = 1;
     void Update()
     {
         transform.Rotate(0, rotateSpeed, 0, Space.World);
